// types/dialog.ts
export interface DialogConfig {
  borderThickness?: number;
  borderColor?: number;
  borderAlpha?: number;
  windowAlpha?: number;
  windowColor?: number;
  windowHeight?: number;
  padding?: number;
  depth?: number;
  closeBtnColor?: string;
  closeBtnFontSize?: number;
  fontSize?: number;
  speed?: number;
}